controller: add support for lighthouse

+++ bot_controller/models.py
@@ -31,9 +31,9 @@ class DotBotRgbLedCommandModel(BaseModel):
 class DotBotLH2Position(BaseModel):
     """Position of a DotBot."""
 
-    x: int
-    y: int
-    z: int
+    x: float
+    y: float
+    z: float
 
 
 class DotBotModel(BaseModel):
